homeworks/homework05 packed classes into directories, renamed classes' methods
