--- II_JavaCoreMiddle/2.1_JavaCore/src/main/java/org/example/IMovable.java
@@ -1,7 +1,0 @@
-package org.example;
-
-public interface IMovable {
-
-    public void Move(double deltaX, double deltaY);
-    public void ChangePosition(double X, double Y);
-}
