+++ II_JavaCoreMiddle/2.1_JavaCore/src/main/java/org/example/figures/Figure.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.figures;
 
 import lombok.Getter;
 import lombok.ToString;
