+++ II_JavaCoreMiddle/2.1_JavaCore/src/main/java/org/example/interfaces/IMovable.java
@@ -1,0 +1,7 @@
+package org.example.interfaces;
+
+public interface IMovable {
+
+    public void move(double deltaX, double deltaY);
+    public void changePosition(double X, double Y);
+}
